Move kwok/cmd to pkg
